- tranferred controllers from admin to mew_administrateur - removed admin/attestation_templates views - changed routes accordingly
