+++ app/views/new_administrateur/attestation_templates/show.pdf.prawn
